flipping through history changes lineup
I rewrote the logic here to avoid anything more
complicated that pushing or poping pages.

I struggled to find algorithm that was symetric
between old and new pages. Both loops now 
preserve the prefix (pages that are matching)
which I moved out of logic and into a variable.

This was expressed as an 'unless' statement but
coffeescript messed up the application of 
demorgan's law.