show how to add things to ipfs
This commit is a very simple example of how to use ipfs to store
assets. We basically just read a stream of data and pass it to
ipfs.add(stream). The response includes the hashes of files added.

(Of course, actually hosting the server-side fed wiki code on ipfs
and doing everything client side is possible. This is just a small
bit to start :)

On backing up the content:

This commit shows how to add content to an ipfs node. the node
will also "pin" the content, which means the node will back it up
and serve it to other nodes. Pins can be removed with:

  ipfs pin rm -r <hash>

On the API Address:

The ipfsApi('localhost', 5001) can be changed to point to whatever
node can be written to. Note that for now, while ipfs is still
figuring out either (a) the way to do a writable HTTP gateway with
low permissions, and (b) the node-ipfs implementation, we need to
setup CORS to work correctly with the server. You should be able to
do:

  ipfs config --json API.HTTPHeaders.Access-Control-Allow-Origin '["localhost:3000", "localhost:5001"]'

(or whatever ports you're using for fedwiki and the ipfs api)
If it doesn't appear to work, you can just run:

  ipfs daemon --unrestricted-api

But please do not run ipfs nodes like this all the time. it exposes
the API to any user.