Fix issues with drag and drop.
Treat remote pages as read-only when dragging. You cannot move items
within them and dragging an item from them is always a copy (never a
move).

Prior to this change remote items could be modified which would result
in the remote page being forked.

Render the lineup starting at the earlier of the original location or
the new location.

When copying or moving an item, the rendering was broken if the copy or
move was to a position more than one spot before the original location.