refactor to introduce internal [{site, slug}...] structure
We introduce an internal structure to represent the Reader's intent
for the lineup. We believe this opens several opportunities.

We enables a URL search parameter to provide the intended lineup while
preserving support for existing use of URL pathname. Using URL search
instead of URL pathname opens the door to statically hosted
wikis. (n.b. introducing this additional behavior means this commit is
not a pure refactoring)

We anticipate this intermediate structure enabling other changes to
clarify responsibilities between state and lineup.

The {site, slug} invites {site, slug, revision} and could open the
door for better support of non-European languages with something like
{site, slug?, revision, title}

Previous experiments in the DAT/hypercore wiki variant and seran
outputs have used the URL hash where this commit chooses to use URL
search.

One rationale is to avoid collisions in our URL semantics with
whatever Google are experimenting with in their recent chrome plugin
for links to text fragments. Google use #:~:... to delimit these new
fragment links.

One other decision hidden in this code is the choice to use the
specific search parameter "pathname" to hold the intended lineup. This
opens experimenting with other URL structures like the
slug@site/slug@site we have in some other experiments.