use location.hash instead of location.search
This commit revisits a decision described in commit 4ea7ab1.
We previously chose location.search partly to avoid collisions with
Google Chrome's experiments with links to text fragments. We have
learned that wiki-server replies with 302 redirects.

In this commit we choose to use location.hash to allow us to release
this functionality without requiring changes to wiki-server.