state.fromLocation() now understands welcome-visitors.html
This commit enables wiki-client to support a statically hosted wiki
while maintaining its support for dynamically hosted node-based wikis.

We hope to find time to follow this commit with some refactoring. In
particular, state.syncDomWithLocation() is hacking the DOM of the html
page early in the bootstrapping of the client, undoing or redoing some
of the work wiki-server does for us. This will unfortunately increase
the confusion in the wiki bootstrapping.

There are details in the client's bootstrapping that need
disentangling between wiki.coffee, legacy.coffee, and state.coffee.
Some changes will also be wanted in wiki-server.

What we get in exchange for this increase in confusion is the ability
for authors to host wikis in github pages or in similar static web
sites. We believe this paves the way toward easier wiki hosting.