clarify names: pathname refers only to url.pathname
We choose "stat" for the url.hash parameter. It gestures at "static"
which matches original intentions which motivate this collection of
changes. Because URLs can get bookmarked or included in links in page
content, we expect to have to support this name for the long term. We
like that it's both short and a bit opaque.