rename parameter from stat to sfw
On returning to this pull request over a year later, the name stat
seems awkward. We now think SFW, the shorthand for the original
Smallest Federated Wiki, is a better name given the origins of this
url schema.

We have also sorted the methods to put fromLocation() and toURL() next
to one another to clarify their inverse relationship. The former
consumes a URL and emits an array of siteSlugs. The latter consumes an
array of siteSlugs and emits a URL.