Add explicit reference to wiki module in files that refer to wiki
I get that this could be a weird design given how wiki is the top
module and then everything else refers to it, but it appears that
the only circular reference is page.js and that is only because
of `siteAdapter.site`.  But referring directly to siteAdapter.site
in page.js fails the window.location.host initialization order
for the Grunt tests.